% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustmrev-package.R
\docType{package}
\name{clustmrev}
\alias{clustmrev-package}
\alias{clustmrev}
\title{clustmrev: A package to review clustering methods applied to mortality}
\description{
The mmcfa package provides ...
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
